+++ R/install.R
@@ -34,6 +34,7 @@ install_keras <- function(method = c("auto", "virtualenv", "conda"),
       conda = conda,
       version = version,
       extra_packages = c("pandas", "Pillow",
+                         "pydot",
                          "tensorflow-hub",
                          "tensorflow-datasets",
                          extra_packages),
@@ -60,14 +61,22 @@ install_keras <- function(method = c("auto", "virtualenv", "conda"),
   )
 }
 
-default_version <- numeric_version("2.8")
+default_version <- numeric_version("2.9")
 
-default_extra_packages <- function(tensorflow_version) {
-  pkgs <- c("tensorflow-hub", "scipy", "requests", "pyyaml", "Pillow", "h5py", "pandas")
+default_extra_packages <- function(tensorflow_version = "default") {
+  pkgs <- c(
+    "tensorflow-hub",
+    "scipy",
+    "requests",
+    "pyyaml",
+    "Pillow",
+    "h5py",
+    "pandas",
+    "pydot")
   names(pkgs) <- pkgs
   v <- tensorflow_version
 
-  if(grepl("nightly", v))
+  if(grepl("nightly|release", v))
     return(pkgs)
 
   ## extract just the version